// src/shared/models/share-user.model.ts
import z from "zod";
import { UserStatus } from "../constant/auth.constant";

export const UserSchema = z.object({
   id: z.number(),
   email: z.string().email(),
   name: z.string().min(1).max(100),
   password: z.string().min(6).max(100),
   phoneNumber: z.string().min(9).max(15),
   avatar: z.string().nullable(),
   totpSecret: z.string().nullable(),
   status: z.enum([UserStatus.ACTIVE, UserStatus.BLOCKED, UserStatus.INACTIVE]),
   roleId: z.number().positive(),
   createdById: z.number().nullable(),
   updatedById: z.number().nullable(),
   deletedAt: z.date().nullable(),
   createdAt: z.date(),
   updatedAt: z.date(),
})

export type UserType = z.infer<typeof UserSchema>;